[SP-277] refactor: jwt 패키지 common으로 이동
