+++ src/main/java/com/cupid/jikting/member/handler/OAuth2LoginSuccessHandler.java
@@ -2,7 +2,7 @@ package com.cupid.jikting.member.handler;
 
 import com.cupid.jikting.common.error.ApplicationError;
 import com.cupid.jikting.common.error.NotFoundException;
-import com.cupid.jikting.jwt.service.JwtService;
+import com.cupid.jikting.common.jwt.service.JwtService;
 import com.cupid.jikting.member.entity.Member;
 import com.cupid.jikting.member.entity.Role;
 import com.cupid.jikting.member.oauth2.CustomOAuth2User;
@@ -35,7 +35,7 @@ public class OAuth2LoginSuccessHandler implements AuthenticationSuccessHandler {
         log.info("OAuth2 Login 성공!");
         CustomOAuth2User oAuth2User = (CustomOAuth2User) authentication.getPrincipal();
         if (oAuth2User.getRole() == Role.GUEST) {
-            String accessToken = jwtService.createAccessToken(getMemberProfileIdByUsername(oAuth2User.getUsername()));
+            String accessToken = jwtService.issueAccessToken(getMemberProfileIdByUsername(oAuth2User.getUsername()));
             response.addHeader(jwtService.getAccessHeader(), TOKEN_TYPE + accessToken);
             response.sendRedirect(OAUTH_SIGNUP_REDIRECT_URL);
             jwtService.sendAccessAndRefreshToken(response, accessToken, null);
@@ -45,8 +45,8 @@ public class OAuth2LoginSuccessHandler implements AuthenticationSuccessHandler {
     }
 
     private void loginSuccess(HttpServletResponse response, CustomOAuth2User oAuth2User) throws IOException {
-        String accessToken = jwtService.createAccessToken(getMemberProfileIdByUsername(oAuth2User.getUsername()));
-        String refreshToken = jwtService.createRefreshToken();
+        String accessToken = jwtService.issueAccessToken(getMemberProfileIdByUsername(oAuth2User.getUsername()));
+        String refreshToken = jwtService.issueRefreshToken();
         response.addHeader(jwtService.getAccessHeader(), "Bearer " + accessToken);
         response.addHeader(jwtService.getRefreshHeader(), "Bearer " + refreshToken);
         jwtService.sendAccessAndRefreshToken(response, accessToken, refreshToken);
