+++ src/test/java/com/cupid/jikting/recommend/controller/RecommendControllerTest.java
@@ -8,7 +8,7 @@ import com.cupid.jikting.common.entity.Personality;
 import com.cupid.jikting.common.error.ApplicationError;
 import com.cupid.jikting.common.error.ApplicationException;
 import com.cupid.jikting.common.error.NotFoundException;
-import com.cupid.jikting.jwt.service.JwtService;
+import com.cupid.jikting.common.jwt.service.JwtService;
 import com.cupid.jikting.member.entity.*;
 import com.cupid.jikting.recommend.dto.MemberResponse;
 import com.cupid.jikting.recommend.dto.RecommendResponse;
@@ -126,7 +126,7 @@ public class RecommendControllerTest extends ApiDocument {
                 .mapToObj(n -> recommendResponse)
                 .collect(Collectors.toList());
         recommendNotFound = new NotFoundException(ApplicationError.RECOMMEND_NOT_FOUND);
-        accessToken = jwtService.createAccessToken(ID);
+        accessToken = jwtService.issueAccessToken(ID);
     }
 
     @WithMockUser
